Add attached files to backups
